+++ app/log.ini
@@ -1,0 +1,38 @@
+[loggers]
+keys=root, spider
+
+[handlers]
+keys=stream_handler, file_handler
+
+[formatters]
+keys=spider_formatter
+
+[logger_root]
+# record all log, only used to root
+level=NOTSET
+handlers=stream_handler
+
+[logger_spider]
+level=DEBUG
+handlers=file_handler
+propagate=1
+qualname=spider
+
+[handler_stream_handler]
+class=StreamHandler
+level=DEBUG
+formatter=spider_formatter
+args=(sys.stderr,)
+
+[handler_file_handler]
+class=FileHandler
+level=DEBUG
+formatter=spider_formatter
+args=('spider.log', 'w')
+
+[formatter_spider_formatter]
+format= "%(asctime)-15s %(levelname)s %(filename)s %(lineno)d %(process)d %(message)s"
+datefmt=
+class=logging.Formatter
+
+# referance to: http://python.usyiyi.cn/documents/python_278/library/logging.config.html#logging-config-api
