+++ tests/test_log.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from app.util.logger import logger
+
+
+logger.error("abc")
